fix some mistakes
apostrophe on can't, added a sentence in the dicision1 and the \n3 in. if that's okay what I added. you must continue the story and add scenarios to it if not it will just be a loop such as "This would be the story continuing after the user's second response" and responding to the third also since I don't know how you want your story